pyproject.toml
[tool.poetry]
name = "invoice-agent"
version = "0.1.0"
description = ""
authors = ["uncleben006 <uncleben006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
motor = "^3.4.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
